+++ src/js/admin.js
@@ -1,8 +1,5 @@
 /**
- * Admin only scripts
+ * Admin only scripts.
  *
  * @package
  */
-console.log(
-	'r3-id-documentation js/admin.js has loaded and has to buy some new hats.'
-);