+++ src/theme-json/styles.mjs
@@ -1,4 +1,6 @@
 export default {
+    // For refrence to settings avalable to theme.json v1: https://developer.bu.edu/gutenberg/best-practices/bu-child-starter/theme-json/
+	// Additional refrence avalable on https://developer.wordpress.org/block-editor/reference-guides/theme-json-reference/theme-json-v1/#styles
     "styles": {
         "elements": {
             "link": {